pyproject.toml
[tool.poetry]
name = "arduino-data-link"
version = "0.1.0"
description = ""
authors = ["GrossBetruger <koriatoren@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pyserial = "^3.5"
numpy = "^2.2.4"
plotly = "^6.0.1"
pandas = "^2.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
